Atualizações: Debug, Leitura da linha no csv

+++ TRAB-02/src/hash/hash.hpp
@@ -47,7 +47,7 @@ struct Diretorio{
         }
         outFile << profundidadeGlobal << endl;
         outFile << maxId << endl;
-        for (int i = 0; i < bucketRefs.size(); i++) {
+        for (size_t i = 0; i < bucketRefs.size(); i++) {
             outFile << bucketRefs[i] << endl;
         }
         outFile.close();
